Add todos and shortcomings/critiques I have for my parser impl
I feel the parser in particular is lacking and where mentioned poorly
structured. I like the flow of assigning A-, C- and Label-Instructions
but feel like the second half of the parser feels awkward to use in
main.
Additionally the main loops work but look and don't feel structured.
Code gen and the symbol table are simple enough and rely mostly on zig
built-in options, so they seem fine to me.